pt1_proyecto_sd/scraper/requirements.txt
requests>=2.25.1
pymongo>=4.0.0
python-dotenv>=0.19.0